chore(analytics): ensure differentiation between marketNews and articles
Previously both "News" and "Artsy Editorial" screens would both report screen
events with an ownerType of `articles`. This commit ensures that we use both
`marketNews` and `articles` owner types appropriately.
