+++ src/app/Scenes/Articles/Articles.tsx
@@ -1,10 +1,15 @@
+import { ActionType, ContextModule, OwnerType } from "@artsy/cohesion"
 import { Screen } from "@artsy/palette-mobile"
+import { ArticleCard_article$data } from "__generated__/ArticleCard_article.graphql"
 import { ArticleSorts, ArticlesQuery } from "__generated__/ArticlesQuery.graphql"
 import { Articles_articlesConnection$key } from "__generated__/Articles_articlesConnection.graphql"
 import { goBack } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
+import { ProvideScreenTrackingWithCohesionSchema } from "app/utils/track"
+import { screen } from "app/utils/track/helpers"
 import React, { Suspense, useState } from "react"
 import { useLazyLoadQuery, usePaginationFragment, graphql } from "react-relay"
+import { useTracking } from "react-tracking"
 import { ArticlesList, ArticlesPlaceholder } from "./ArticlesList"
 
 export const Articles: React.FC = () => {
@@ -16,6 +21,7 @@ export const Articles: React.FC = () => {
   >(articlesConnectionFragment, queryData)
 
   const [refreshing, setRefreshing] = useState(false)
+  const tracking = useTracking()
 
   const handleLoadMore = () => {
     if (!hasNext || isLoadingNext) {
@@ -31,22 +37,34 @@ export const Articles: React.FC = () => {
     setRefreshing(false)
   }
 
+  const handleOnPress = (article: ArticleCard_article$data) => {
+    const tapEvent = tracks.tapArticlesListItem(article.internalID, article.slug || "")
+    tracking.trackEvent(tapEvent)
+  }
+
   const articles = extractNodes(data.articlesConnection)
 
   return (
     <Screen>
-      <Screen.AnimatedHeader onBack={goBack} title="Artsy Editorial" />
-      <Screen.StickySubHeader title="Artsy Editorial" />
-      <Screen.Body fullwidth>
-        <ArticlesList
-          articles={articles as any}
-          isLoading={() => isLoadingNext}
-          hasMore={() => hasNext}
-          refreshing={refreshing}
-          handleLoadMore={handleLoadMore}
-          handleRefresh={handleRefresh}
-        />
-      </Screen.Body>
+      <ProvideScreenTrackingWithCohesionSchema
+        info={screen({
+          context_screen_owner_type: OwnerType.articles,
+        })}
+      >
+        <Screen.AnimatedHeader onBack={goBack} title="Artsy Editorial" />
+        <Screen.StickySubHeader title="Artsy Editorial" />
+        <Screen.Body fullwidth>
+          <ArticlesList
+            articles={articles as any}
+            isLoading={() => isLoadingNext}
+            hasMore={() => hasNext}
+            refreshing={refreshing}
+            handleLoadMore={handleLoadMore}
+            handleRefresh={handleRefresh}
+            handleOnPress={handleOnPress}
+          />
+        </Screen.Body>
+      </ProvideScreenTrackingWithCohesionSchema>
     </Screen>
   )
 }
@@ -94,3 +112,14 @@ const articlesConnectionFragment = graphql`
     }
   }
 `
+
+export const tracks = {
+  tapArticlesListItem: (articleId: string, articleSlug: string) => ({
+    action: ActionType.tappedArticleGroup,
+    context_module: ContextModule.articles,
+    context_screen_owner_type: OwnerType.articles,
+    destination_screen_owner_type: OwnerType.article,
+    destination_screen_owner_id: articleId,
+    destination_screen_owner_slug: articleSlug,
+  }),
+}
